Backend: Updated auth

+++ backend/src/controllers/post/post.controller.js
@@ -13,9 +13,18 @@ exports.create = async (req, res, next) => {
   }
 
   try {
-    const newPostFromDatabase = await postService.create(req.body);
-    await blogService.addPost(newPostFromDatabase.author.username, newPostFromDatabase._id);
-    return res.status(201).json(newPostFromDatabase);
+    const user = await userService.findOneById(req.body.author);
+    if (!user) {
+      return next(new createError.NotFound('Author not found'));
+    }
+
+    if (req.user && req.user.email && req.user.email === user.email) {
+      const newPostFromDatabase = await postService.create(req.body);
+      await blogService.addPost(newPostFromDatabase.author.username, newPostFromDatabase._id);
+      return res.status(201).json(newPostFromDatabase);
+    }
+
+    return next(new createError.Unauthorized());
   } catch (err) {
     return next(new createError.InternalServerError(err.message));
   }
@@ -50,8 +59,18 @@ exports.update = async (req, res, next) => {
       return next(new createError.NotFound('Post not found'));
     }
 
-    const updatedPost = await postService.update(req.params.id, req.body);
-    return res.json(updatedPost);
+    if (req.user && req.user.email) {
+      const user = await userService.findOneByEmail(req.user.email);
+      if (!user) {
+        return next(new createError.Unauthorized());
+      }
+
+      if ((req.user.email === post.author.email) || (user.role === 'admin' && post.visibility === 'public')) {
+        const updatedPost = await postService.update(req.params.id, req.body);
+        return res.json(updatedPost);
+      }
+    }
+    return next(new createError.Unauthorized());
   } catch (err) {
     return next(new createError.InternalServerError(err.message));
   }
@@ -64,9 +83,19 @@ exports.delete = async (req, res, next) => {
       return next(new createError.NotFound('Post not found'));
     }
 
-    await postService.delete(req.params.id);
-    await blogService.deletePost(post.author.username, req.params.id);
-    return res.json({});
+    if (req.user && req.user.email) {
+      const user = await userService.findOneByEmail(req.user.email);
+      if (!user) {
+        return next(new createError.Unauthorized());
+      }
+
+      if ((req.user.email === post.author.email) || (user.role === 'admin' && post.visibility === 'public')) {
+        await postService.delete(req.params.id);
+        await blogService.deletePost(post.author.username, req.params.id);
+        return res.json({});
+      }
+    }
+    return next(new createError.Unauthorized());
   } catch (err) {
     return next(new createError.InternalServerError(err.message));
   }
